Fix some gce-only tests to run on gke as well
DaemonRestart kubelet test, and Services apiserver restart test.
